requirements.txt
altgraph==0.15
astroid==1.6.0
backports.functools-lru-cache==1.4
configparser==3.5.0
dis3==0.1.2
enum34==1.1.6
future==0.16.0
isort==4.2.15
lazy-object-proxy==1.3.1
macholib==1.9
mccabe==0.6.1
pefile==2017.11.5
Pillow==5.0.0
PyInstaller==3.3.1
pylint==1.8.1
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
